fightv14.5 (NOT WORKING)
Deleting the enemy from the array would shift the enemy positions within the array and cause the incorrect enemy to be removed from the array, hence causing an enemy attack to be executed while said enemy does not exist within the array, causing a softlock. This wip update is attempting to instead replace the enemy with a string rather than completely deleting it, therefore not causing the aforementioned desynchronization.