shop system finished
just a few things need to be polished but it's fully functional, minus the sell mechanic. again, next step is to make these items functional in battle, as well as create a system that makes it so that the items stack in the player's inventory. or maybe not. I might save that for later, or just not do it. makes it harder for the player